--- uastreamer/src/comps/mod.rs
@@ -1,4 +1,0 @@
-pub mod tcp;
-pub mod udp;
-pub mod audio;
-pub mod cli;
