+++ uastreamer/src/pooled/cpal.rs
@@ -39,79 +39,6 @@ pub struct CpalStats {
     pub output_info: Option<OutputCallbackInfo>,
 }
 
-pub fn select_input_device_config(
-    device: &cpal::Device,
-    requested_bufsize: u32,
-    requested_samplerate: u32,
-    chcount: usize,
-) -> SupportedStreamConfig {
-    let devs = device.supported_input_configs().unwrap().find(|x| {
-        let bsize = *x.buffer_size();
-        let bsize_match = match bsize {
-            cpal::SupportedBufferSize::Range { min, max } => {
-                min <= requested_bufsize && requested_bufsize <= max
-            }
-            cpal::SupportedBufferSize::Unknown => false,
-        };
-        
-        debug!(
-            "min_sample_rate: {}, max_sample_rate: {}",
-            x.min_sample_rate().0,
-            x.max_sample_rate().0
-        );
-
-        return x.min_sample_rate().0 <= requested_samplerate
-            && x.max_sample_rate().0 >= requested_samplerate
-            && x.channels() >= chcount as u16;
-    });
-
-    println!("{:?}", devs);
-
-    match devs {
-        Some(d) => d
-            .with_sample_rate(cpal::SampleRate(requested_samplerate)),
-        None => device
-            .default_input_config()
-            .expect("failed loading default input config")
-    }
-}
-
-pub fn select_output_device_config(
-    device: &cpal::Device,
-    requested_bufsize: u32,
-    requested_samplerate: u32,
-    chcount: usize,
-) -> SupportedStreamConfig {
-    let devs = device.supported_output_configs().unwrap().find(|x| {
-        let bsize = *x.buffer_size();
-        let bsize_match = match bsize {
-            cpal::SupportedBufferSize::Range { min, max } => {
-                min <= requested_bufsize && requested_bufsize <= max
-            }
-            cpal::SupportedBufferSize::Unknown => false,
-        };
-
-        debug!(
-            "min_sample_rate: {}, max_sample_rate: {}",
-            x.min_sample_rate().0,
-            x.max_sample_rate().0
-        );
-
-        return x.min_sample_rate().0 <= requested_samplerate
-            && x.max_sample_rate().0 >= requested_samplerate
-            && x.channels() >= chcount as u16;
-    });
-
-    println!("{:?}", devs);
-
-    match devs {
-        Some(d) => d
-            .with_sample_rate(cpal::SampleRate(requested_samplerate)),
-        None => device
-            .default_output_config()
-            .expect("failed loading default output config"),
-    }
-}
 
 #[deprecated]
 pub trait CpalAudioFlow<T: cpal::SizedSample + Send + Pod + Default + Debug + 'static> {
